Cache black configuration per file
Use lru_cache decorator on load_config to remember the black configuration
for each file that the plugin processes. If the black configuration
changes for any reason (e.g. the user editing the pyproject.toml file for
the project) then the LSP server would need to be restarted for
python-lsp-black to refresh the configuration.